% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_selective.R
\name{drop_selective}
\alias{drop_selective}
\title{Drop and slice with controlled dimensionality reduction}
\usage{
drop_selective(x, dim_extract, dim_drop)
}
\arguments{
\item{x}{A multi-dimensional array}

\item{dim_extract}{Which level in `dim_drop` to extract}

\item{dim_drop}{Index of the dimension to slice and drop.}
}
\description{
Drop and slice with controlled dimensionality reduction
}
\examples{
my_array <- array(NA, dim = c(3, 1, 5, 10))
# We want to take the first out of ten levels in the 4th dimension
# and then drop this (and only this) dimension.
my_array[,,,1] \%>\% dim()
# This reduces the array to two dimensions!
my_array[,,,1, drop = FALSE] \%>\% dim()
# This keeps all the four dimensions!
drop_selective(my_array, 1, 4) \%>\% dim()
# Ta-daa!
}
\author{
Michal Oleszak
}
