% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hridge.R
\name{get_varbased_weights}
\alias{get_varbased_weights}
\title{Produce variance-based weights for predictors in heteroskedastic ridge regression}
\usage{
get_varbased_weights(y, X)
}
\arguments{
\item{y}{Vector of response variable}

\item{X}{Matrix of predictor variables}
}
\value{
A vector of length ncol(X) with weights for each predictor.
}
\description{
Produce variance-based weights for predictors in heteroskedastic ridge regression
}
