% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hridge.R
\name{hridge}
\alias{hridge}
\title{Heteroskedastic Ridge Regression}
\usage{
hridge(y, X, lambda, predictor_weights = "varbased",
  optim_method = "Nelder-Mead")
}
\arguments{
\item{y}{Vector of response variable}

\item{X}{Matrix of predictor variables}

\item{lambda}{Value of ridge (L2) regularization penalty}

\item{predictor_weights}{Vector of length ncol(X) with weights for predictor variables.
Defaults to "varbased", in which case weights are based on the variance of parameter
estimates obtained through a set of univaraite regressions.}

\item{optim_method}{Optimization algorithm, passed to optim()}
}
\value{
A list with parameter estimates, fitted values and multiple R-squared.
}
\description{
Heteroskedastic Ridge Regression
}
\author{
Michal Oleszak
}
