% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hridge.R
\name{hridge_loss}
\alias{hridge_loss}
\title{Loss Function in Heteroskedastic Ridge Regression}
\usage{
hridge_loss(betas)
}
\arguments{
\item{betas}{vector of parameter estimates}
}
\value{
A single numeric, value of the loss function.
}
\description{
Loss Function in Heteroskedastic Ridge Regression
}
\author{
Michal Oleszak
}
