% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_robust_aggregators.R
\name{mean_}
\alias{mean_}
\title{Arithmetic mean of values ignoring NAa}
\usage{
mean_(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A single numeric.
}
\description{
Arithmetic mean of values ignoring NAa
}
\examples{
a <- c(1, 2, 3, NA, 5)
mean_(a)
}
