% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_robust_aggregators.R
\name{median_}
\alias{median_}
\title{Median of values ignoring NAs}
\usage{
median_(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A single numeric.
}
\description{
Median of values ignoring NAs
}
\examples{
a <- c(1, 2, 3, NA, 5)
median_(a)
}
