% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_robust_aggregators.R
\name{min_}
\alias{min_}
\title{Minimal value ignoring NAa}
\usage{
min_(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A single numeric.
}
\description{
Minimal value ignoring NAa
}
\examples{
a <- c(1, 2, 3, NA, 5)
min_(a)
}
