% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_robust_aggregators.R
\name{sum_}
\alias{sum_}
\title{Sum of values ignoring NAa}
\usage{
sum_(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A single numeric.
}
\description{
Sum of values ignoring NAa
}
\examples{
a <- c(1, 2, 3, NA, 5)
sum_(a)
}
